+++ custom_components/volcano_hybrid/__init__.py
@@ -16,6 +16,7 @@ PLATFORMS: list[Platform] = [
     Platform.NUMBER,
     Platform.SENSOR,
     Platform.SWITCH,
+    Platform.UPDATE,
 ]